Move dataclass tests into test37 folder
