+++ test/test_attr.py
@@ -1,6 +1,3 @@
-from dataclasses import dataclass
-from typing import ClassVar
-
 import inject
 from test import BaseTestInject
 
@@ -29,27 +26,8 @@ class TestInjectAttr(BaseTestInject):
         assert value0 == 123
         assert value1 == 123
 
-    def test_attr_on_dataclass_class_var_raises_error(self):
-        with self.assertRaises(inject.InjectorException):
-            @dataclass
-            class MyClass:
-                field: ClassVar[int] = inject.attr(int)
-
 
 class TestInjectAttrDataclass(BaseTestInject):
-    def test_attr_dc(self):
-        @dataclass
-        class MyClass(object):
-            field = inject.attr_dc(int)
-
-        inject.configure(lambda binder: binder.bind(int, 123))
-        my = MyClass()
-        value0 = my.field
-        value1 = my.field
-
-        assert value0 == 123
-        assert value1 == 123
-
     def test_class_attr_dc(self):
         class MyClass(object):
             field = inject.attr_dc(int)
@@ -61,12 +39,3 @@ class TestInjectAttrDataclass(BaseTestInject):
         assert value0 == 123
         assert value1 == 123
 
-    def test_attr_on_dataclass_class_var_works(self):
-        @dataclass
-        class MyClass:
-            field: ClassVar[int] = inject.attr_dc(int)
-
-        inject.configure(lambda binder: binder.bind(int, 123))
-
-        assert MyClass().field == 123
-        assert MyClass.field == 123
